LayoutTests/fast/dom/beforeload/cached-image-before-load-expected.txt

This is a regression test for http://webkit.org/b/45586. It verifies that the beforeload event fires before the load event when loading a cached image, and that the image's style attributes can be queried in the load event handler. On success, you should see two lines of green text starting with the word 'PASS'.

PASS: load fired after beforeload for a cached image.
PASS: image is 16px wide.
 
